pyproject.toml
[tool.poetry]
name = "FortiFy"
version = "0.1.0"
description = "App para acompanhar a rotina de treinos"
authors = ["Luiz <lcsouzagarcia@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
kivy = "^2.3.1"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
redis = "^5.2.1"
asyncpg = "^0.30.0"
argon2-cffi = "^23.1.0"
pyjwt = "^2.10.1"
psycopg2-binary = "^2.9.10"
fastapi-mail = "^1.4.2"

 
[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"
ruff = "^0.8.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]
format = "ruff format"
lint = "ruff check"
run_test = "fastapi dev api/main.py --app MAIN_APP"


[tool.ruff.lint]
select = ["I", "E", "F", "W", "C"]